requirements.txt
pytest
allure-pytest
selenium
requests
mysql-connector-python
pytest-html